eliminate UnitOps
